requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
atproto==0.0.56
cloudscraper==1.2.71
dnspython==2.7.0
pymongo==4.10.1
python-dotenv==1.0.1